refactor and add try-catch block
