+++ src/components/member-profile/contribution/index.js
@@ -6,6 +6,7 @@ import PRLink from '@components/member-profile/contribution/pr-link';
 import { timeWas } from '@helper-functions/time-was';
 import { HOST_NAME } from '@constants/AppConstants';
 import { useState } from 'react';
+import ContributionCollapsedCard from '@components/member-profile/contribution/ContributionCollapsedCard';
 import SuperUserOptions from '../super-user-options/container';
 
 const renderPRLinks = (prList) =>
@@ -177,37 +178,6 @@ const ContributionCard = ({
           {renderFeatureUrl(url, urlObj)}
         </div>
       )}
-      <hr className={classNames.line} />
-    </div>
-  );
-};
-const ContributionCollapsedCard = ({ contribution }) => {
-  const {
-    task: { id, title, isNoteworthy, isCollapsed },
-    prList,
-  } = contribution;
-  const isTitleAvailable = !!title;
-  const featureTitle = isTitleAvailable ? title : prList[0].title;
-  const [showSettings, setShowSettings] = useState(false);
-
-  return (
-    <div
-      onMouseEnter={() => setShowSettings(true)}
-      onMouseLeave={() => setShowSettings(false)}
-    >
-      <div className={classNames.contributionContainer}>
-        <div className={classNames.leftSection}>
-          <h3 className={classNames.featureTitle}>{featureTitle}</h3>
-        </div>
-        <div className={classNames.rightSection}>
-          <SuperUserOptions
-            showSettings={showSettings}
-            isNoteworthy={isNoteworthy}
-            isCollapsed={isCollapsed}
-            taskId={id}
-          />
-        </div>
-      </div>
       <hr className={classNames.line} />
     </div>
   );
